+++ NEWS.md
@@ -1,5 +1,6 @@
 # dsem 1.4.2
 
+* Added `total_effect(.)` to calculate total effects
 * Fixed bug where dsem crashed when `quiet = FALSE` and running without data
   e.g., as qualitative network model